Make config files be found case-insensitively
By using the <lowercasePackageName>-config.cmake and
<lowercasePackageName>-config-version.cmake naming schemes for the
generated CMake config files, we make it possible for these config files
to be found, irrespective of the capitalization used in the original
find_package call. That is, find_package(NaMe) and find_package(name)
will both find the same config file.

This is already the case no case-insensitive filesystems (e.g. under
Windows). However, on case-sensitive filesystems (e.g. most Linux
setups) the used capitalization makes a difference. This can be a source
of confusion and annoyance in cross-platform development.

With this commit, this platform differences will vanish.