Increase config file compat w.r.t different capitalization
The config file will now detect wrong capitalization of the package name
and emit a warning about it. Furthermore, it will try its best to
subsequently use the package name in the capitalization as used in the
respective find_package call. This should minimize the issues that can
arise from wrong capitalization (such as mismatched capitalization in
the resulting variables).
Nonetheless, there are likely some issues remaining, which is why the
warning requests fixing the capitalization of the package name.